backend/requirements.txt
flask
flask-cors
beautifulsoup4
python-dotenv
sqlalchemy
psycopg2-binary
waitress
requests
celery
redis
google-generativeai
playwright
playwright-stealth
boto3
2captcha-python
tweepy